requirements.txt
Flask==2.3.2
selenium==4.20.0
pandas==2.0.3
Unipath==1.1
mysql-connector